+++ src/highscore.js
@@ -1,13 +1,26 @@
 /** @jsx jsx */
+import { useState, useEffect } from 'react';
+import useRemoteData from './useRemoteData';
 import { Link } from '@reach/router';
 import { jsx } from '@emotion/core';
+import Loading from './loading';
 
 function Highscore() {
+	const [ highscore, setHighscore ] = useState([]);
+
+	const { data, loadingState } = useRemoteData('http://localhost:5555/api/highscore');
+	useEffect( () => {
+		if( loadingState === 'loaded' ) {
+			setHighscore( data );
+		}
+	}, [ data, loadingState, setHighscore ]);
+
 	return (
-		<div>
+		<Loading data={ highscore } loadingState={ loadingState }>
 			Higscore yay!
 			<Link to='/'>Go back</Link>
-		</div>
+			data: { JSON.stringify(highscore) }
+		</Loading>
 	);
 };
